requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
future==0.17.1
gspread==3.1.0
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
oauth2client==4.1.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dotenv==0.10.2
pytz==2019.1
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.3
Werkzeug==0.15.2
